Move `FeedImageDataLoaderPresentationAdapter` to separate file
